requirements.txt
torch>=2.0.0
opencv-python
diffusers==0.25.1
accelerate
peft==0.7.1
transformers>=4.35.1
safetensors
controlnet-aux==0.0.3
fastapi==0.108.0
uvicorn
python-multipart
python-socketio==5.7.2
typer
pydantic>=2.5.2
rich
loguru
yacs
piexif==1.1.3
omegaconf
easydict
gradio
typer-config==1.4.0

Pillow==9.5.0 # for AnyText
